requirements.txt
selenium
pytest
webdriver-manager
openpyxl
requests
pyyaml
Appium-Python-Client
pytest-html